+++ src/views/layout.js
@@ -206,7 +206,7 @@ function renderLayout({ title, activePath = '', body = '', script = '' } = {}) {
           ${renderNavLinks('footer', activePath)}
           <button class="logout-button" onclick="window.location.href='/logout'">Abmelden</button>
           <div class="sidebar-copyright">
-            Copyright (C) 2026 Kevin Käfer | MyKaefer Apps<br>
+            Copyright (C) 2026 Kevin Käfer | <a class="sidebar-copyright-link" href="https://apps.mykaefer.net" target="_blank" rel="noopener noreferrer">MyKaefer Apps</a><br>
             Version: ${escapeHtml(pkgVersion)}
           </div>
         </div>